chore: setup.yml

+++ zsh/.zshrc
@@ -1,7 +1,7 @@
-export JAVA_HOME=/usr/lib/jvm/default
 export NODE_ENV=development
 export CHROME_EXECUTABLE=/opt/google/chrome/chrome
-export PATH=/home/vic/.cargo/bin:$PATH
+export PATH=$HOME/.cargo/bin:$PATH
+export PATH=$HOME/.local/bin:$PATH
 export PATH=~/.dotnet/tools:$PATH
 export PATH=~/go/bin:$PATH
 export ANDROID_HOME=$HOME/Android/Sdk
@@ -20,34 +20,46 @@ setopt INC_APPEND_HISTORY_TIME
 setopt EXTENDED_HISTORY
 
 bindkey -e
-zstyle :compinstall filename '/home/vic/.zshrc'
+zstyle :compinstall filename "$HOME/.zshrc"
 
 autoload -Uz compinit
 compinit
 
-source /usr/share/zsh/plugins/zsh-autosuggestions/zsh-autosuggestions.zsh
-source /usr/share/zsh/plugins/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh
+for _plugin_path in \
+    /usr/share/zsh/plugins/zsh-autosuggestions/zsh-autosuggestions.zsh \
+    /usr/share/zsh-autosuggestions/zsh-autosuggestions.zsh; do
+    [[ -f $_plugin_path ]] && source $_plugin_path && break
+done
 
-__conda_setup="$('/opt/anaconda/bin/conda' 'shell.bash' 'hook' 2> /dev/null)"
-if [ $? -eq 0 ]; then
-    eval "$__conda_setup"
-else
-    if [ -f "/opt/anaconda/etc/profile.d/conda.sh" ]; then
-        . "/opt/anaconda/etc/profile.d/conda.sh"
+for _plugin_path in \
+    /usr/share/zsh/plugins/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh \
+    /usr/share/zsh-syntax-highlighting/zsh-syntax-highlighting.zsh; do
+    [[ -f $_plugin_path ]] && source $_plugin_path && break
+done
+unset _plugin_path
+
+if [[ -d /opt/anaconda ]]; then
+    __conda_setup="$('/opt/anaconda/bin/conda' 'shell.bash' 'hook' 2> /dev/null)"
+    if [ $? -eq 0 ]; then
+        eval "$__conda_setup"
     else
-        export PATH="/opt/anaconda/bin:$PATH"
+        if [ -f "/opt/anaconda/etc/profile.d/conda.sh" ]; then
+            . "/opt/anaconda/etc/profile.d/conda.sh"
+        else
+            export PATH="/opt/anaconda/bin:$PATH"
+        fi
     fi
+    unset __conda_setup
 fi
-unset __conda_setup
 
 function mkcd {
-        if [ ! -n "$1" ]; then
-                echo "Enter a directory name"
-        elif [ -d "$1" ]; then
-                echo "\$1 already exists"
-        else
-                mkdir -p $1 && cd $1
-        fi
+    if [ ! -n "$1" ]; then
+        echo "Enter a directory name"
+    elif [ -d "$1" ]; then
+        echo "\$1 already exists"
+    else
+        mkdir -p "$1" && cd "$1"
+    fi
 }
 
 export NVM_DIR="$HOME/.nvm"
@@ -55,12 +67,12 @@ export NVM_DIR="$HOME/.nvm"
 [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"
 
 function nvm() {
-  unfunction nvm
-  \. "$NVM_DIR/nvm.sh"
-  nvm "$@"
-  if [[ "$1" == "use" || "$1" == "install" ]]; then
-    corepack enable 2>/dev/null
-  fi
+    unfunction nvm
+    \. "$NVM_DIR/nvm.sh"
+    nvm "$@"
+    if [[ "$1" == "use" || "$1" == "install" ]]; then
+        corepack enable 2>/dev/null
+    fi
 }
 
 alias vim=nvim
@@ -74,46 +86,49 @@ export FZF_DEFAULT_COMMAND="fd --hidden --strip-cwd-prefix --exclude .git"
 export FZF_CTRL_T_COMMAND="$FZF_DEFAULT_COMMAND"
 export FZF_ALT_C_COMMAND="fd --type=d --hidden --strip-cwd-prefix --exclude .git"
 _fzf_compgen_path() {
-  fd --hidden --exclude .git . "$1"
+    fd --hidden --exclude .git . "$1"
 }
 _fzf_compgen_dir() {
-  fd --type=d --hidden --exclude .git . "$1"
+    fd --type=d --hidden --exclude .git . "$1"
 }
-source ~/fzf-git.sh/fzf-git.sh
+[[ -f ~/fzf-git.sh/fzf-git.sh ]] && source ~/fzf-git.sh/fzf-git.sh
 export FZF_CTRL_T_OPTS="--preview 'bat -n --color=always --line-range :500 {}'"
 export FZF_ALT_C_OPTS="--preview 'eza --tree --color=always {} | head -200'"
 
 _fzf_comprun() {
-  local command=$1
-  shift
-
-  case "$command" in
-    cd)           fzf --preview 'eza --tree --color=always {} | head -200' "$@" ;;
-    export|unset) fzf --preview "eval 'echo \$'{}"         "$@" ;;
-    ssh)          fzf --preview 'dig {}'                   "$@" ;;
-    *)            fzf --preview "bat -n --color=always --line-range :500 {}" "$@" ;;
-  esac
+    local command=$1
+    shift
+    case "$command" in
+        cd)           fzf --preview 'eza --tree --color=always {} | head -200' "$@" ;;
+        export|unset) fzf --preview "eval 'echo \$'{}"         "$@" ;;
+        ssh)          fzf --preview 'dig {}'                   "$@" ;;
+        *)            fzf --preview "bat -n --color=always --line-range :500 {}" "$@" ;;
+    esac
 }
 
-alias drop-caches='sudo paccache -rk3; yay -Sc --aur --noconfirm'
-alias update-all='export TMPFILE="$(mktemp)"; \
-    sudo true; \
-    rate-mirrors --save=$TMPFILE arch --max-delay=21600 \
-      && sudo mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist-backup \
-      && sudo mv $TMPFILE /etc/pacman.d/mirrorlist \
-      && drop-caches \
-      && yay -Syyu --noconfirm'
+if [[ -f /etc/arch-release ]]; then
+    alias drop-caches='sudo paccache -rk3; yay -Sc --aur --noconfirm'
+    alias update-all='export TMPFILE="$(mktemp)"; \
+        sudo true; \
+        rate-mirrors --save=$TMPFILE arch --max-delay=21600 \
+          && sudo mv /etc/pacman.d/mirrorlist /etc/pacman.d/mirrorlist-backup \
+          && sudo mv $TMPFILE /etc/pacman.d/mirrorlist \
+          && drop-caches \
+          && yay -Syyu --noconfirm'
+fi
 
 eval "$(fzf --zsh)"
 
 eval "$(zoxide init zsh)"
 alias cd="z"
 
-if [ -f ~/.local/state/quickshell/user/generated/terminal/sequences.txt ]; then
+if [[ -n "$HYPRLAND_INSTANCE_SIGNATURE" ]] \
+   && [[ -f ~/.local/state/quickshell/user/generated/terminal/sequences.txt ]]; then
     cat ~/.local/state/quickshell/user/generated/terminal/sequences.txt
 fi
 
 if command -v wt >/dev/null 2>&1; then eval "$(command wt config shell init zsh)"; fi
-export PATH=/home/vic/.opencode/bin:$PATH
+
+[[ -d $HOME/.opencode/bin ]] && export PATH=$HOME/.opencode/bin:$PATH
 
 eval "$(starship init zsh)"
